calculator/pyproject.toml
[project]
name = "calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
